refactor : booleangenerator 패키지 삭제
BooleanGenerator, RandomBooleanGenerator
domain 패키지로 이동
